[refactor] atomic 패턴에 맞게 폴더 구조 변경
- fetures 하위로 이동
